requirements.txt
streamlit
google-generativeai
requests
textblob
plotly
pyyaml
